+++ plugins/woocommerce-admin/packages/components/src/search-list-control/icons/index.js
@@ -1,5 +1,0 @@
-// Export each icon as a named component.
-export { default as IconCheckChecked } from './checkbox-checked';
-export { default as IconCheckUnchecked } from './checkbox-unchecked';
-export { default as IconRadioSelected } from './radio-selected';
-export { default as IconRadioUnselected } from './radio-unselected';